requirements.txt
python_dateutil==2.5.3
Django==1.9.6
djangorestframework==3.4.1
django_mptt==0.8.5
django_autocomplete_light==3.1.6
requests==2.5.3
tmdbsimple==1.3.0
beautifulsoup4==4.5.0
pillow==3.3.1